support use hdwallet key to sign message in 'ioctl account sign' command
