+++ ioctl/cmd/account/accountsign.go
@@ -51,9 +51,15 @@ func init() {
 }
 
 func accountSign(msg string) error {
-	addr, err := util.GetAddress(signer)
-	if err != nil {
-		return output.NewError(output.InputError, "failed to get signer addr", err)
+	var addr string
+	var err error
+	if !util.AliasIsHdwalletKey(signer) {
+		addr, err = util.GetAddress(signer)
+		if err != nil {
+			return output.NewError(output.InputError, "failed to get signer addr", err)
+		}
+	} else {
+		addr = signer
 	}
 	fmt.Printf("Enter password #%s:\n", addr)
 	password, err := util.ReadSecretFromStdin()
