+++ ioctl/cmd/action/action.go
@@ -11,7 +11,6 @@ import (
 	"encoding/hex"
 	"fmt"
 	"math/big"
-	"strconv"
 	"strings"
 
 	"github.com/golang/protobuf/proto"
@@ -110,37 +109,10 @@ func decodeBytecode() ([]byte, error) {
 	return hex.DecodeString(util.TrimHexPrefix(bytecodeFlag.Value().(string)))
 }
 
-func parseHdwPath(addressOrAlias string) (uint32, uint32, uint32, error) {
-	// parse derive path
-	// for hdw::1/1/2, return 1, 1, 2
-	// for hdw::1/2, treat as default account = 0, return 0, 1, 2
-	args := strings.Split(addressOrAlias[5:], "/")
-	if len(args) < 2 || len(args) > 3 {
-		return 0, 0, 0, output.NewError(output.ValidationError, "derivation path error", nil)
-	}
-
-	arg := make([]uint32, 3)
-	for i := len(args) - 1; i >= 0; i-- {
-		u64, err := strconv.ParseUint(args[i], 10, 32)
-		if err != nil {
-			return 0, 0, 0, output.NewError(output.InputError, fmt.Sprintf("%v must be integer value", args[i]), err)
-		}
-		arg[i] = uint32(u64)
-	}
-	return arg[0], arg[1], arg[2], nil
-}
-
-func aliasIsHdwalletKey(addressOrAlias string) bool {
-	if strings.HasPrefix(strings.ToLower(addressOrAlias), "hdw::") {
-		return true
-	}
-	return false
-}
-
 // Signer returns signer's address
 func Signer() (address string, err error) {
 	addressOrAlias := signerFlag.Value().(string)
-	if aliasIsHdwalletKey(addressOrAlias) {
+	if util.AliasIsHdwalletKey(addressOrAlias) {
 		return addressOrAlias, nil
 	}
 
@@ -154,7 +126,7 @@ func Signer() (address string, err error) {
 }
 
 func nonce(executor string) (uint64, error) {
-	if aliasIsHdwalletKey(executor) {
+	if util.AliasIsHdwalletKey(executor) {
 		// for hdwallet key, get the nonce in SendAction()
 		return 0, nil
 	}
@@ -287,7 +259,7 @@ func SendAction(elp action.Envelope, signer string) error {
 	var prvKey crypto.PrivateKey
 	var err error
 
-	if account.IsSignerExist(signer) || aliasIsHdwalletKey(signer) {
+	if account.IsSignerExist(signer) || util.AliasIsHdwalletKey(signer) {
 		// Get signer's password
 		password := passwordFlag.Value().(string)
 		if password == "" {
@@ -298,8 +270,8 @@ func SendAction(elp action.Envelope, signer string) error {
 			}
 		}
 
-		if aliasIsHdwalletKey(signer) {
-			account, change, index, err := parseHdwPath(signer)
+		if util.AliasIsHdwalletKey(signer) {
+			account, change, index, err := util.ParseHdwPath(signer)
 			if err != nil {
 				return output.NewError(output.InputError, "invalid hdwallet key format", err)
 			}
